# crypto/sha: implement SHA1 & SHA256 acceleration using Intel SHA extensions