--- a/src/internal/cpu/cpu.go
+++ b/src/internal/cpu/cpu.go
@@ -41,6 +41,7 @@ var X86 struct {
 	HasSSSE3     bool
 	HasSSE41     bool
 	HasSSE42     bool
+	HasSHA       bool
 	_            CacheLinePad
 }