This change implements SHA1 & SHA256 acceleration using the Intel SHA
extensions if those instructions are enabled. First the internal/cpu
package needs support to detect the CPU extension flag for SHA extensions.

The crypto/sha256 package contains a small refactor so that it matches how
crypto/sha1 & crypto/sha512 are set up and it makes it easier to use
consistent code in how it's decided which implementation can be used based
on CPU support.

All the work here is based on the Intel reference documentation at
https://software.intel.com/content/www/us/en/develop/articles/intel-sha-extensions.html.
Using only the reference documentation avoids issue like in #27443 where
code from another project was converted to Go assembly.

Benchmarks show a close to 2x performance improvement on an AMD Ryzen 5
3600 for SHA1:
    
name                       old time/op  new time/op  delta
Sha/SHA1______16_bytes-12   172ns ± 3%   105ns ± 4%  -39.06%  (p=0.000 n=20+21)
Sha/SHA1______64_bytes-12   261ns ± 2%   139ns ± 3%  -46.85%  (p=0.000 n=21+21)
Sha/SHA1_____256_bytes-12   492ns ± 2%   229ns ± 2%  -53.41%  (p=0.000 n=20+20)
Sha/SHA1______1k_bytes-12  1.17µs ± 1%  0.59µs ± 1%  -49.43%  (p=0.000 n=20+19)
Sha/SHA1______8k_bytes-12  7.48µs ± 1%  3.94µs ± 1%  -47.35%  (p=0.000 n=20+21)
Sha/SHA1____256k_bytes-12   232µs ± 2%   122µs ± 1%  -47.25%  (p=0.000 n=21+20)
Sha/SHA1___1024k_bytes-12   928µs ± 2%   491µs ± 1%  -47.12%  (p=0.000 n=21+21)

Benchmarks show a close to 4x performance improvement on an AMD Ryzen 5
3600, especially on larger inputs. Even on the smallest it's at least 2x
faster.

name                       old time/op  new time/op  delta
Sha/SHA256____16_bytes-12   248ns ± 3%   117ns ± 3%  -52.84%  (p=0.000 n=20+19)
Sha/SHA256____64_bytes-12   384ns ± 2%   153ns ± 3%  -60.10%  (p=0.000 n=20+17)
Sha/SHA256___256_bytes-12   786ns ± 1%   249ns ± 3%  -68.29%  (p=0.000 n=19+19)
Sha/SHA256____1k_bytes-12  2.36µs ± 1%  0.64µs ± 3%  -72.93%  (p=0.000 n=19+20)
Sha/SHA256____8k_bytes-12  17.0µs ± 2%   4.2µs ± 1%  -75.16%  (p=0.000 n=20+20)
Sha/SHA256__256k_bytes-12   537µs ± 1%   131µs ± 1%  -75.60%  (p=0.000 n=20+20)
Sha/SHA256_1024k_bytes-12  2.15ms ± 1%  0.52ms ± 1%  -75.60%  (p=0.000 n=20+20)

The discussion in #27443 mentions that for SHA1 including this was
debatable, since the algorithm itself is no longer considered safe. I think
that a 2x performance improvement is still significant though and there's
still a lot of places where SHA1 is still used (for example for Git itself).
Of course the SHA1 change can be backed out if this change is only desired
for SHA256 because of that.

Fixes #27443