@dbussink 2023-04-01
Closing as an implementation for SHA256 was submitted by Intel.